chore: split device host devices into one file per device type
