+++ devicehost/device.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/evcc-io/evcc/api"
 	"github.com/evcc-io/evcc/api/implement"
 	"github.com/evcc-io/evcc/devicehost/proto/pb"
 	"github.com/evcc-io/evcc/util"
@@ -90,66 +89,4 @@ func (d *device) decorate(c implement.Caps) {
 			f(d, c)
 		}
 	}
-}
-
-// Meter creates a meter from a device host
-func Meter(ctx context.Context, other map[string]any) (api.Meter, error) {
-	d, err := newDevice(ctx, templates.Meter, other)
-	if err != nil {
-		return nil, err
-	}
-
-	res := &meter{device: d, Caps: implement.New()}
-	d.decorate(res.Caps)
-
-	return res, nil
-}
-
-type meter struct {
-	*device
-	implement.Caps
-}
-
-func (m *meter) CurrentPower() (float64, error) {
-	var power float64
-	err := m.call("api.Meter", "CurrentPower", nil, &power)
-	return power, err
-}
-
-// Charger creates a charger from a device host
-func Charger(ctx context.Context, other map[string]any) (api.Charger, error) {
-	d, err := newDevice(ctx, templates.Charger, other)
-	if err != nil {
-		return nil, err
-	}
-
-	res := &charger{device: d, Caps: implement.New()}
-	d.decorate(res.Caps)
-
-	return res, nil
-}
-
-type charger struct {
-	*device
-	implement.Caps
-}
-
-func (c *charger) Status() (api.ChargeStatus, error) {
-	var status api.ChargeStatus
-	err := c.call("api.ChargeState", "Status", nil, &status)
-	return status, err
-}
-
-func (c *charger) Enabled() (bool, error) {
-	var enabled bool
-	err := c.call("api.Charger", "Enabled", nil, &enabled)
-	return enabled, err
-}
-
-func (c *charger) Enable(enable bool) error {
-	return c.call("api.Charger", "Enable", []any{enable})
-}
-
-func (c *charger) MaxCurrent(current int64) error {
-	return c.call("api.CurrentController", "MaxCurrent", []any{current})
 }
